+++ lib/airbrake/capistrano/tasks.rb
@@ -1,5 +1,0 @@
-require 'airbrake/capistrano'
-
-warn "DEPRECATION WARNING: Requiring 'airbrake/capistrano/tasks' is " \
-     "deprecated and will be removed in the next MAJOR release. Require " \
-     "'airbrake/capistrano' instead."