+++ lib/airbrake/rake/task_ext.rb
@@ -1,5 +1,0 @@
-require 'airbrake/rake'
-
-warn "DEPRECATION WARNING: Requiring 'airbrake/rake/task_ext' is " \
-     "deprecated and will be removed in the next MAJOR release. Require " \
-     "'airbrake/rake' instead."