+++ lib/airbrake/resque/failure.rb
@@ -1,5 +1,0 @@
-require 'airbrake/resque'
-
-warn "DEPRECATION WARNING: Requiring 'airbrake/resque/failure' is " \
-     "deprecated and will be removed in the next MAJOR release. Require " \
-     "'airbrake/resque' instead."