+++ lib/airbrake/shoryuken/error_handler.rb
@@ -1,5 +1,0 @@
-require 'airbrake/shoryuken'
-
-warn "DEPRECATION WARNING: Requiring 'airbrake/shoryuken/error_handler' is " \
-     "deprecated and will be removed in the next MAJOR release. Require " \
-     "'airbrake/shoryuken' instead."