+++ lib/airbrake/sidekiq/error_handler.rb
@@ -1,5 +1,0 @@
-require 'airbrake/sidekiq'
-
-warn "DEPRECATION WARNING: Requiring 'airbrake/sidekiq/error_handler' is " \
-     "deprecated and will be removed in the next MAJOR release. Require " \
-     "'airbrake/sidekiq' instead."